feat(plugin): adapter plugin -- h2
