+++ configure/etc/conf/plugins.properties
@@ -7,6 +7,7 @@ datacap-plugin-alioss=plugin/datacap-plugin-alioss/pom.xml
 datacap-plugin-zookeeper=plugin/datacap-plugin-zookeeper/pom.xml
 datacap-plugin-kafka=plugin/datacap-plugin-kafka/pom.xml
 datacap-plugin-redis=plugin/datacap-plugin-redis/pom.xml
+datacap-plugin-h2=plugin/datacap-plugin-h2/pom.xml
 datacap-convert-txt=convert/datacap-convert-txt/pom.xml
 datacap-convert-json=convert/datacap-convert-json/pom.xml
 datacap-convert-xml=convert/datacap-convert-xml/pom.xml
