+++ core/datacap-ui/public/static/images/plugin/h2.png
