--- plugin/datacap-native-h2/src/main/resources/META-INF/services/io.edurt.datacap.spi.PluginModule
@@ -1,1 +1,0 @@
-io.edurt.datacap.plugin.natived.h2.H2PluginModule
