--- plugin/datacap-native-h2/src/test/java/io/edurt/datacap/plugin/natived/h2/H2PluginTest.java
@@ -1,44 +1,0 @@
-package io.edurt.datacap.plugin.natived.h2;
-
-import com.google.inject.Guice;
-import com.google.inject.Injector;
-import com.google.inject.Key;
-import com.google.inject.TypeLiteral;
-import io.edurt.datacap.spi.PluginService;
-import io.edurt.datacap.spi.model.Configure;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-
-import java.util.Optional;
-import java.util.Set;
-
-public class H2PluginTest
-{
-    private Injector injector;
-    private Configure configure;
-
-    @Before
-    public void before()
-    {
-        injector = Guice.createInjector(new H2PluginModule());
-        configure = new Configure();
-        configure.setHost("-");
-        configure.setPort(0);
-    }
-
-    @Test
-    public void test()
-    {
-        Set<PluginService> plugins = injector.getInstance(Key.get(new TypeLiteral<Set<PluginService>>() {}));
-        Optional<PluginService> pluginOptional = plugins.stream()
-                .filter(v -> v.name().equalsIgnoreCase("H2"))
-                .findFirst();
-        if (pluginOptional.isPresent()) {
-            PluginService plugin = pluginOptional.get();
-            plugin.connect(configure);
-            Assert.assertNotNull(plugin.execute(plugin.validator()).getConnection());
-            plugin.destroy();
-        }
-    }
-}
