+++ plugin/datacap-plugin-h2/src/main/java/io/edurt/datacap/plugin/H2Plugin.java
@@ -1,0 +1,7 @@
+package io.edurt.datacap.plugin;
+
+import io.edurt.datacap.plugin.Plugin;
+
+public class H2Plugin
+        extends Plugin
+{}
